% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.quantile_model}
\alias{predict.quantile_model}
\title{Predict from a \code{quantile_model}}
\usage{
\method{predict}{quantile_model}(object, new_data, ...)
}
\arguments{
\item{object}{A \code{quantile_model} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in \code{new_data}.
}
\description{
Predict from a \code{quantile_model}
}
\examples{
predictors <- mtcars[, -1]
outcome <- mtcars[, 1]
# TODO Add some examples

}
