% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{quantile_model}
\alias{quantile_model}
\alias{quantile_model.default}
\alias{quantile_model.data.frame}
\alias{quantile_model.matrix}
\alias{quantile_model.formula}
\alias{quantile_model.recipe}
\title{Fit a \code{quantile_model}}
\usage{
quantile_model(x, ...)

\method{quantile_model}{default}(x, ...)

\method{quantile_model}{data.frame}(x, y, quantiles, ...)

\method{quantile_model}{matrix}(x, y, quantiles, ...)

\method{quantile_model}{formula}(formula, data, quantiles, ...)

\method{quantile_model}{recipe}(x, data, quantiles, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{...}{Not currently used, but required for extensibility.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 numeric column.
\item A \strong{matrix} with 1 numeric column.
\item A numeric \strong{vector}.
}}

\item{quantiles}{A numeric vector of \emph{at least} one quantile (>0 and <1).}

\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
}}
}
\value{
A \code{quantile_model} object.
}
\description{
\code{quantile_model()} fits a model.
}
\examples{
predictors <- mtcars[, -1]
outcome <- mtcars[, 1]
# TODO Add some examples

}
